Use NPX to build an client
NPX is a tool that its goal is to help round out the experience of using packages from the NPM registry. As NPM makes it super easy to install and manage dependencies hosted on the registry, NPX makes it easy to use CLI tools and other executables hosted on the registry.

Knowing this, let’s create our director. If you have the server side for the Backend, we have the client side for the Frontend.

$ npx create-react-app client
$ cd client
$ yarn start
$ brew install yarn
Add PORT=8001 to scripts in package.json